1.2.0
Uniapp_Frontend_Final

+++ Tsicket-uniapp/pages/user/user.vue
@@ -1,75 +1,6 @@
 <template>
-	<!-- <view class="flex-page">
-		<view class="userinfo">
-			<image class="usercard" src="../../static/cardback0.jpg"></image>
-			<view v-if="hasUserInfo">
-				<view class="padding flex align-start justify-between">
-					<view class="cu-avatar xxl round" :style="{backgroundImage: userInfo.avatarUrl ? 'url(' + userInfo.avatarUrl + ')' : ''}"></view>
-					<view class="text-white">
-						<view class="text-right">
-							<view class="text-sl text-ellipsis">
-								<text class="text-bold">{{userInfo.nickName}}</text>
-							</view>
-						</view>
-						<view class="text-right">
-							<button v-if="!hasTsinghuaInfo" class="cu-btn round light bg-blue" @click="identification">清华身份认证</button>
-							<text v-else>{{tsinghuaid}}</text>
-						</view>
-						<view class="text-left flex justify-start text-sm padding">
-							<view class="padding-right-xl" @click="likePage">
-								<view class="text-xl text-bold">{{like}}</view>
-								喜爱
-							</view>
-							<view class="padding-right-xl" @click="followPage">
-								<view class="text-xl text-bold">{{follow}}</view>
-								关注
-							</view>
-							<view class="padding-right-xl" @click="historyPage">
-								<view class="text-xl text-bold">{{history}}</view>
-								历史
-							</view>
-						</view>
-					</view>
-				</view>
-			</view>
-			<view v-else class="flex justify-center getinfobtn">
-				<button class="cu-btn round" open-type="getUserInfo" @getuserinfo="getUserInfo"> 授权个人信息</button>
-			</view>
-		</view>
-		<view class="tabs">
-			<view class="nav" :current="current" @change="navChange">
-				<scroll-view scroll-x class="bg-white nav">
-					<view class="flex text-center">
-						<view class="flex-sub tab" :class="index==current?'tab-choose':''" v-for="(item,index) in tabs" :key="index" @tap="tabSelect"
-						 :data-id="index">
-							{{item}}
-						</view>
-					</view>
-				</scroll-view>
-			</view>
-			<view class="tab-swiper-view">
-				<swiper class="tab-swiper" :current="current" @change="swiperChange">
-					<swiper-item>
-						<scroll-view scroll-y class="tab-scroll padding">
-							<view class="flex-column">
-							<activity-prepare v-for="(item, index) in userActivity" :key="item.event_id" :activity="item" @clickSponsor="sponsorPage"
-							 @detail="activityPage(item.event_id)" @announcement="announcementPage" @QRcode="QRcodePage" @cancel="cancelActivity(index)" @like="likeActivity(index)"
-							 @unfold="unfold(index)"
-							 :class="[item.delay?'animation-slide-bottom': item.cancel]"
-							  :style="[{animationDelay: item.delay}]"></activity-prepare>
-							 </view>
-						</scroll-view>
-					</swiper-item>
-					<swiper-item>
-						<scroll-view scroll-y class="tab-scroll padding">
-							<activity-check :activity="activity" :sponsor="sponsor" :message="message" @like="like" @clickCard="activityPage"></activity-check>
-						</scroll-view>
-					</swiper-item>
-				</swiper>
-			</view>
-		</view>
-	</view> -->
 	<view>
+		<!-- class="flex-page"> -->
 		<view class="userinfo flex-column">
 			<image class="usercard" src="../../static/cardback0.jpg"></image>
 			<view v-if="hasUserInfo">
@@ -235,9 +166,9 @@
 		methods: {
 			loadpage() {
 				uni.request({
-					url: app.globalData.apiurl + 'users/view', //仅为示例，并非真实接口地址。
+					url: app.globalData.apiurl + 'users/view',
 					header: {
-						'content-type': 'application/json', //自定义请求头信息
+						'content-type': 'application/json',
 						'cookie': app.globalData.cookie
 					},
 					success: (res) => {
@@ -255,9 +186,9 @@
 				})
 
 				uni.request({
-					url: app.globalData.apiurl + 'users/book', //仅为示例，并非真实接口地址。
+					url: app.globalData.apiurl + 'users/book',
 					header: {
-						'content-type': 'application/json', //自定义请求头信息
+						'content-type': 'application/json',
 						'cookie': app.globalData.cookie
 					},
 					success: (res) => {
@@ -267,8 +198,8 @@
 							res.isUnfold = false
 							res.delay = '' + (index + 1) * 0.1 + 's'
 							res.cancel = undefined
-							if(!res.img_url || (res.img_url == '') ) {
-								res.img_url =  app.globalData.backimg[parseInt('11' + res.event_id) % 4]
+							if (!res.img_url || (res.img_url == '')) {
+								res.img_url = app.globalData.backimg[parseInt('11' + res.event_id) % 4]
 							}
 							setTimeout(() => {
 								res.delay = undefined
@@ -322,9 +253,9 @@
 			},
 			historyPage() {
 				console.log("historypage")
-				// uni.navigateTo({
-				// 	url: "../history/history"
-				// })
+				uni.navigateTo({
+					url: "../history/history"
+				})
 			},
 			sponsorPage(name) {
 				uni.navigateTo({
@@ -343,24 +274,44 @@
 
 			},
 			cancelActivity(index) {
-				uni.request({
-					url: app.globalData.apiurl + 'users/book/' + this.userActivity[index].event_id, //仅为示例，并非真实接口地址。
-					method: 'DELETE',
-					header: {
-						'content-type': 'application/json', //自定义请求头信息
-						'cookie': app.globalData.cookie
-					},
+				uni.showModal({
+					title: '',
+					content: '请确认是否取消参加\r\n' + this.userActivity[index].event_name,
 					success: (res) => {
-						console.log(res);
-						this.userActivity[index].delay = undefined
-						this.userActivity[index].cancel = 'animation-delete-slide-left'
-						setTimeout(() => {
-							this.userActivity[index].isUnfold = false
-							this.userActivity.splice(index, 1)
-						}, 2000)
-					},
-					fail: (res) => {
-
+						if (res.confirm) {
+							uni.request({
+								url: app.globalData.apiurl + 'users/book/' + this.userActivity[index].event_id,
+								method: 'DELETE',
+								header: {
+									'content-type': 'application/json',
+									'cookie': app.globalData.cookie
+								},
+								success: (res) => {
+									console.log(res);
+									if (res.data.success) {
+										this.isreserved = false
+										this.userActivity[index].delay = undefined
+										this.userActivity[index].cancel = 'animation-delete-slide-left'
+										setTimeout(() => {
+											this.userActivity[index].isUnfold = false
+											this.userActivity.splice(index, 1)
+										}, 2000)
+										uni.showToast({
+											icon: 'none',
+											title: '取消成功'
+										})
+									} else {
+										uni.showToast({
+											icon: 'none',
+											title: '操作失败'
+										})
+									}
+								},
+								fail: (res) => {
+									console.log(res)
+								}
+							})
+						}
 					}
 				})
 			},
@@ -369,7 +320,7 @@
 					url: app.globalData.apiurl + 'users/like/' + this.userActivity[index].event_id,
 					method: 'POST',
 					header: {
-						'content-type': 'application/json', //自定义请求头信息
+						'content-type': 'application/json',
 						'cookie': app.globalData.cookie
 					},
 					success: (res) => {
@@ -392,7 +343,6 @@
 						type: "id.tsinghua"
 					},
 					success(res) {
-						// 打开成功
 						console.log(res)
 					},
 					fail(res) {
